Tracer monitoring tests drive a scripted client instead of a real model
